requirements.txt
rich>=13.0.0
psutil>=5.8.0
Pillow>=9.0.0
numpy>=1.20.0
cryptography>=3.4.0
pycryptodome>=3.15.0
